requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
apipkg==2.1.0
appdirs==1.4.4
async-timeout==4.0.0
asynctest==0.13.0
atomicwrites==1.4.0
attrs==21.2.0
beautifulsoup4==4.10.0
browserstack-local==1.2.3
cached-property==1.5.2
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.7
colorama==0.3.7
cycler==0.11.0
defusedxml==0.7.1
docopt==0.6.2
et-xmlfile==1.1.0
execnet==1.5.0
frozenlist==1.2.0
glob2==0.7
google-api-python-client==1.6.4
httplib2==0.10.3
idna==2.6
imageio==2.10.4
importlib-metadata==4.8.1
imutils==0.5.4
isodate==0.6.0
kiwisolver==1.3.2
lxml==4.6.4
Mako==1.1.5
MarkupSafe==2.0.1
matplotlib==3.4.3
more-itertools==8.10.0
multidict==5.2.0
networkx==2.6.3
num2words==0.5.10
numpy==1.21.4
oauth2client==4.1.3
opencv-python==4.5.4.58
openpyxl==3.0.9
packaging==21.2
pandas==1.3.4
parse==1.19.0
parse-type==0.5.2
Pillow==8.4.0
pluggy==0.13.1
psutil==5.8.0
py==1.5.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoIt==0.6.5
pyparsing==2.4.7
PyPDF2==1.26.0
pytest==5.4.3
pytest-bdd==5.0.0
pytest-forked==1.3.0
pytest-xdist==1.34.0
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
requests==2.18.4
requests-toolbelt==0.9.1
rsa==4.7.2
sauceclient==1.0.0
scikit-image==0.18.3
scipy==1.7.2
selenium==3.8.0
six==1.16.0
soupsieve==2.3
tifffile==2021.11.2
tinydb==3.14.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.22
wcwidth==0.2.5
xlrd==1.1.0
yarl==1.7.2
zeep==2.4.0
zipp==3.6.0
